Add initial implementation of sending AgeSDLHook state on link-in
This is uncommitted code from two months ago. I'm relatively sure that
it doesn't work properly in this state, because the SDL blob from the
vault has to be adjusted in *some* way by the server so that the client
handles it correctly. But this implementation doesn't break anything
either and it should be a good start for proper SDL handling on the
server side.