Explicitly use sphinx_rtd_theme for documentation
This used to be configured implicitly as part of the Read the Docs
build, but they stopped doing this a while ago, so the theme reverted to
Sphinx' default Alabaster theme. Local builds have also always used
Alabaster.

This change restores the original behavior on Read the Docs and makes
local builds consistent with that.