Revert toctrees from :titlesonly: back to :maxdepth: 1
This seems to be necessary for sphinx_rtd_theme to generate working
section navigation in the sidebar. With :titlesonly:, the sidebar only
shows page titles and never the sections within the pages.

This doesn't make a big difference. The only disadvantage is that the
game server subpages now take an extra click to reach from the main
index page.